requirements.txt
telethon
httpx
python-dotenv
rich
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
pystray
Pillow
